DX-836 Percentage Fix with Dashboard Components
